Bye bye ColorIconButton👋 (& other random stuff)
- Removed ColorIconButton, in favor or MUI's IconButton (with custom global styling)
- MarkdownInputBase buttons now display differently depending on screen size and left-handedness
- Fixed bug where "Enter" button didn't work for MarkdownInputBase. This was caused by FindInPage's hotkey
