+++ packages/shared/src/translations/locales/en/common.json
@@ -11,7 +11,9 @@
     "AddAccount": "Add account",
     "AddFocusMode": "Add focus mode",
     "AddNew": "Add New",
+    "AddNewLink": "Add new link",
     "AddApi": "Add API",
+    "AddBookmark": "Add bookmark",
     "AddComment": "Add comment",
     "AddNote": "Add note",
     "AddOrganization": "Add team",
@@ -22,6 +24,7 @@
     "AddSmartContract": "Add smart contract",
     "AddStandard": "Add standard",
     "AddThisDevice": "Add this device",
+    "AddToTeam": "Add to team",
     "AddWallet": "Add wallet",
     "AdvancedSearch": "Advanced Search",
     "AdditionalResources": "Additional Resources",
@@ -72,6 +75,7 @@
     "ChangePassword": "Change Password",
     "ChangePasswordBeforeLogin": "Before signing in, please follow the link sent to your email to change your password.",
     "Chat": "Chat",
+    "CleanGraph": "Clean up graph",
     "ClearSearchHistory": "Clear search history",
     "Close": "Close",
     "Code": "Code",
@@ -128,7 +132,9 @@
     "CreateBot": "Create bot",
     "CreateBotDescription": "Create a bot to chat with and use in your routines",
     "CreateChat": "Create chat",
+    "CreateEvent": "Create event",
     "CreateFocusMode": "Create focus mode",
+    "CreateInvite": "Create invite",
     "CreateMeeting": "Create meeting",
     "CreateNote": "Create note",
     "CreateNoteDescription": "Write down your thoughts, ideas, & more",
@@ -231,6 +237,7 @@
     "Faq": "FAQ",
     "Feature": "Feature",
     "Feed": "Feed",
+    "FilterList": "Filter list",
     "Find": "Find",
     "FindInPage": "Find in page...",
     "FindUsOn": "Find us on...",
@@ -382,6 +389,7 @@
     "Message_other": "Messages",
     "MessagesAsc": "Messages ↑",
     "MessagesDesc": "Messages ↓",
+    "MessageSend": "Send message",
     "MicrophonePermissionDenied": "Permission denied. Please allow access to your microphone.",
     "Min": "Min",
     "MinRows": "Min rows",
@@ -784,6 +792,8 @@
     "TagSelectorPlaceholder": "Entar tags, followed by commas...",
     "Targeting": "Targeting",
     "Terms": "Terms",
+    "TextGrow": "Make text larger",
+    "TextShrink": "Make text smaller",
     "TextSize": "Text size",
     "Theme": "Theme",
     "ThisIsMultiStep": "This is a multi-step routine.",
@@ -820,6 +830,7 @@
     "UpdateBot": "Update bot",
     "UpdateChat": "Update chat",
     "UpdateFocusMode": "Update focus mode",
+    "UpdateList": "Update list",
     "UpdateMeeting": "Update meeting",
     "UpdateNote": "Update note",
     "UpdateOrganization": "Update team",
