+++ packages/ui/src/components/buttons/ColorIconButton/ColorIconButton.tsx
@@ -1,22 +1,11 @@
 import { IconButton } from "@mui/material";
 import { forwardRef } from "react";
-import { ColorIconButtonProps } from "../types";
-
-export const buttonSx = (background: string, disabled: boolean | undefined) => ({
-    background,
-    pointerEvents: disabled ? "none" : "auto",
-    filter: disabled ? "grayscale(1) opacity(0.5)" : "none",
-    transition: "filter 0.2s ease-in-out",
-    "&:hover": {
-        background,
-        filter: disabled ? "grayscale(1) opacity(0.5)" : "brightness(1.2)",
-    },
-} as const);
+import { highlightStyle } from "styles";
 
 /**
  * IconButton with a custom color
  */
-export const ColorIconButton = forwardRef<HTMLButtonElement | HTMLAnchorElement, ColorIconButtonProps>(({
+export const ColorIconButton = forwardRef<HTMLButtonElement | HTMLAnchorElement, any>(({
     background,
     children,
     disabled,
@@ -32,7 +21,7 @@ export const ColorIconButton = forwardRef<HTMLButtonElement | HTMLAnchorElement,
             href={href}
             onClick={onClick}
             sx={{
-                ...buttonSx(background, disabled),
+                ...highlightStyle(background, disabled),
                 ...sx,
             }}
         >
@@ -45,7 +34,7 @@ export const ColorIconButton = forwardRef<HTMLButtonElement | HTMLAnchorElement,
             {...props}
             onClick={onClick}
             sx={{
-                ...buttonSx(background, disabled),
+                ...highlightStyle(background, disabled),
                 ...sx,
             }}
         >
