+++ packages/ui/src/components/buttons/relationships/QuestionForButton/QuestionForButton.tsx
@@ -1,6 +1,5 @@
 import { exists, QuestionForType } from "@local/shared";
 import { IconButton, Stack, Tooltip, Typography, useTheme } from "@mui/material";
-import { buttonSx } from "components/buttons/ColorIconButton/ColorIconButton";
 import { FindObjectDialog } from "components/dialogs/FindObjectDialog/FindObjectDialog";
 import { SelectOrCreateObjectType } from "components/dialogs/types";
 import { RelationshipItemQuestionForObject } from "components/lists/types";
@@ -9,6 +8,7 @@ import { AddIcon, ApiIcon, NoteIcon, OrganizationIcon, ProjectIcon, RoutineIcon,
 import { useCallback, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useLocation } from "route";
+import { highlightStyle } from "styles";
 import { SvgComponent } from "types";
 import { getDisplay } from "utils/display/listTools";
 import { openObject } from "utils/navigation/openObject";
@@ -117,7 +117,7 @@ export function QuestionForButton({
                             borderRadius: 8,
                             paddingRight: 2,
                             ...largeButtonProps(isEditing, true),
-                            ...buttonSx(palette.primary.light, !isEditing),
+                            ...highlightStyle(palette.primary.light, !isEditing),
                         }}
                     >
                         <IconButton>
