+++ packages/ui/src/views/MemberManageView/MemberManageView.tsx
@@ -1,6 +1,5 @@
-import { CommonKey, MemberInviteStatus } from "@local/shared";
+import { CommonKey } from "@local/shared";
 import { IconButton, Stack, Tooltip, Typography, useTheme } from "@mui/material";
-import { ColorIconButton } from "components/buttons/ColorIconButton/ColorIconButton";
 import { SideActionButtons } from "components/buttons/SideActionButtons/SideActionButtons";
 import { MaybeLargeDialog } from "components/dialogs/LargeDialog/LargeDialog";
 import { SearchList } from "components/lists/SearchList/SearchList";
@@ -12,20 +11,8 @@ import { AddIcon, LockIcon, SearchIcon, UnlockIcon } from "icons";
 import { useCallback, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { toDisplay } from "utils/display/pageTools";
-import { MemberManagePageTabOption, SearchType } from "utils/search/objectToSearch";
+import { MemberManagePageTabOption, memberTabParams } from "utils/search/objectToSearch";
 import { MemberManageViewProps } from "../types";
-
-const tabParams = [{
-    titleKey: "Member" as CommonKey,
-    searchType: SearchType.Member,
-    tabType: MemberManagePageTabOption.Member,
-    where: (organizationId: string) => ({ organizationId }),
-}, {
-    titleKey: "Invite" as CommonKey,
-    searchType: SearchType.MemberInvite,
-    tabType: MemberManagePageTabOption.MemberInvite,
-    where: (organizationId: string) => ({ organizationId, statuses: [MemberInviteStatus.Pending, MemberInviteStatus.Declined] }),
-}];
 
 /**
  * View members and invited members of an organization
@@ -45,7 +32,7 @@ export const MemberManageView = ({
         searchType,
         tabs,
         where,
-    } = useTabs<MemberManagePageTabOption>({ tabParams, display });
+    } = useTabs<MemberManagePageTabOption>({ tabParams: memberTabParams, display });
 
     const [isOpenToNewMembersField, , isOpenToNewMembersHelpers] = useField("isOpenToNewMembers");
 
@@ -130,12 +117,12 @@ export const MemberManageView = ({
                 display={display}
                 sx={{ position: "fixed" }}
             >
-                <ColorIconButton aria-label="filter-list" background={palette.secondary.main} onClick={focusSearch} >
+                <IconButton aria-label={t("FilterList")} onClick={focusSearch} sx={{ background: palette.secondary.main }}>
                     <SearchIcon fill={palette.secondary.contrastText} width='36px' height='36px' />
-                </ColorIconButton>
-                <ColorIconButton aria-label="edit-routine" background={palette.secondary.main} onClick={onInviteStart} >
+                </IconButton>
+                <IconButton aria-label={t("CreateInvite")} onClick={onInviteStart} sx={{ background: palette.secondary.main }}>
                     <AddIcon fill={palette.secondary.contrastText} width='36px' height='36px' />
-                </ColorIconButton>
+                </IconButton>
             </SideActionButtons>
         </MaybeLargeDialog>
     );
